+++ tests/NumberTest.php
@@ -5,20 +5,25 @@ declare(strict_types=1);
 namespace Serhii\Tests;
 
 use PHPUnit\Framework\Attributes\DataProvider;
+use Serhii\ShortNumber\Config;
 use Serhii\ShortNumber\Exceptions\LargeNumberException;
 use Serhii\ShortNumber\Number;
 
 final class NumberTest extends TestCase
 {
     #[DataProvider('provideNumberShort')]
-    public function testNumberShort(string $expect, int $num): void
+    public function testNumberShort(string $expect, int $num, Config|null $config = null): void
     {
+        Number::configure($config ?? new Config());
         $this->assertSame($expect, Number::short($num));
     }
 
     public static function provideNumberShort(): array
     {
         return self::withNegativeNumbers([
+            ['5.9m', 5_935_235, new Config(decimals: 1)],
+            ['8.135m', 8_135_235, new Config(decimals: 3)],
+            ['1__10m', 1_105_000, new Config(decimals: 2, decimalsSeparator: '__')],
             ['0', 0],
             ['1', 1],
             ['500', 500],